CLN: clean benchmarks to get them running (#16025)
* fix lib and algos import

* fix take_1d import

* string uppercase -> ascii_uppercase (py3 compat)

* sas test file path

* fix datetools usage

* fix hashing benchmarks

* dict values py3 compat

* avoid overflow by using higher freq

* xrange -> range

* fix xport path

* revised hdfstore_bench to use new query syntax
rename table variables

* change default python version to 3.6

+++ asv_bench/benchmarks/groupby.py
@@ -331,7 +331,7 @@ class groupby_multi(object):
 
     def get_test_data(self, ngroups=100, n=100000):
         self.unique_groups = range(self.ngroups)
-        self.arr = np.asarray(np.tile(self.unique_groups, (n / self.ngroups)), dtype=object)
+        self.arr = np.asarray(np.tile(self.unique_groups, int(n / self.ngroups)), dtype=object)
         if (len(self.arr) < n):
             self.arr = np.asarray((list(self.arr) + self.unique_groups[:(n - len(self.arr))]), dtype=object)
         random.shuffle(self.arr)
